+++ libs/cardano-ledger-api/src/Cardano/Ledger/Api/Tx.hs
@@ -34,7 +34,7 @@ module Cardano.Ledger.Api.Tx (
   module Cardano.Ledger.Api.Tx.Wits,
 
   -- * Shelley onwards
-  EraTx,
+  EraTx (Tx),
   mkBasicTx,
   bodyTxL,
   witsTxL,