+++ eras/allegra/impl/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Version history for `cardano-ledger-allegra`
 
-## 1.0.0.1
+## 1.1.0.0
+
+* Remove redundant pattern synonym `AllegraTxAuxData'`
+* Hide internal `AllegraTxAuxDataRaw` constructor with `atadrMetadata` and `atadrTimelock`
+  record fields.
 
 ### `testlib`