+++ eras/alonzo/impl/CHANGELOG.md
@@ -9,6 +9,7 @@
 * Move `Cardano.Ledger.Alonzo.Tools` module into `cardano-ledger-api:Cardano.Ledger.Api.Scripts`
 * Add helper lens `hashDataTxWitsL`
 * Rename `smMap` to `cmValues`
+* Remove redundant pattern synonym `AlonzoTxAuxData'{atadMetadata',atadTimelock',atadPlutus'}`
 
 ###`testlib`